Move all service interfaces into interfaces folder
