+++ src/interfaces.rs
@@ -1,12 +1,9 @@
+pub mod block;
 pub mod messenger;
+pub mod packet_processor;
+pub mod patchwork;
+pub mod player;
 
 use super::models::map;
-use super::models::minecraft_types;
 use super::models::packet;
 use super::models::translation;
-
-use super::packet_handlers;
-
-use super::packet_handlers::packet_router;
-
-use super::server;
