+++ src/packet_handlers.rs
@@ -9,8 +9,6 @@ pub mod initiation_protocols;
 pub mod packet_router;
 pub mod peer_subscription;
 
-use super::services::game_state;
-
 use super::models::packet;
 use super::models::translation;
