+++ src/server.rs
@@ -1,5 +1,5 @@
 use super::interfaces::messenger::Messenger;
-use super::services::packet_processor::PacketProcessor;
+use super::interfaces::packet_processor::PacketProcessor;
 
 use super::models::minecraft_protocol::MinecraftProtocolReader;
