+++ src/services.rs
@@ -17,9 +17,6 @@ use super::models::packet;
 use super::models::translation;
 
 use super::interfaces;
-
 use super::packet_handlers;
-
 use super::packet_handlers::packet_router;
-
 use super::server;
