+++ src/services/game_state/patchwork.rs
@@ -1,61 +1,20 @@
 use super::gameplay_router;
 use super::interfaces::messenger::Messenger;
+use super::interfaces::packet_processor::PacketProcessor;
+use super::interfaces::patchwork::PatchworkStateOperations;
+use super::interfaces::player::PlayerState;
 use super::map::{Map, Peer, Position};
 use super::packet;
 use super::packet::Packet;
-use super::packet_processor::PacketProcessor;
-use super::player::PlayerState;
 use super::server;
 use std::collections::HashMap;
 use std::io;
 use std::sync::mpsc::Receiver;
-use std::sync::mpsc::Sender;
+
 use uuid::Uuid;
 
 pub const ENTITY_ID_BLOCK_SIZE: i32 = 1000;
 pub const CHUNK_SIZE: i32 = 16;
-
-pub trait PatchworkState {
-    fn new_map(&self, peer: Peer);
-    fn route_player_packet(&self, packet: Packet, conn_id: Uuid);
-    fn report(&self);
-}
-
-impl PatchworkState for Sender<PatchworkStateOperations> {
-    fn new_map(&self, peer: Peer) {
-        self.send(PatchworkStateOperations::New(NewMapMessage { peer }))
-            .unwrap();
-    }
-
-    fn route_player_packet(&self, packet: Packet, conn_id: Uuid) {
-        self.send(PatchworkStateOperations::RoutePlayerPacket(RouteMessage {
-            packet,
-            conn_id,
-        }))
-        .unwrap();
-    }
-
-    fn report(&self) {
-        self.send(PatchworkStateOperations::Report).unwrap();
-    }
-}
-
-pub enum PatchworkStateOperations {
-    New(NewMapMessage),
-    RoutePlayerPacket(RouteMessage),
-    Report,
-}
-
-#[derive(Debug)]
-pub struct NewMapMessage {
-    pub peer: Peer,
-}
-
-#[derive(Debug, Clone)]
-pub struct RouteMessage {
-    pub packet: Packet,
-    pub conn_id: Uuid,
-}
 
 pub fn start<
     M: 'static + Messenger + Clone + Send,
